[Common] Refactor ECU File Organizer structure
This update restructures the ECU File Organizer project by separating its components into individual modules within an 'ecu_file_organizer' package. Key changes include the creation of '__init__.py', '__main__.py', 'constants.py', 'form_dialog.py', 'main_window.py', 'monitor.py', and 'parser.py' to better organize the codebase. This modular approach enhances maintainability, readability, and scalability of the application. Additionally, a 'requirements.txt' file was added to explicitly manage project dependencies.
