+++ ecu_file_organizer/__init__.py
@@ -1,0 +1,5 @@
+"""ECU File Organizer - Automatic ECU file organization for automotive diagnostics."""
+
+from ecu_file_organizer.constants import APP_VERSION
+
+__version__ = APP_VERSION
