+++ ecu_file_organizer/constants.py
@@ -1,0 +1,18 @@
+"""Shared constants for ECU File Organizer."""
+
+APP_VERSION = "1.7"
+APP_NAME = "ECU File Organizer"
+APP_DISPLAY_NAME = f"ECU File Organizer v{APP_VERSION}"
+SUPPORT_URL = "https://buymeacoffee.com/autobyte"
+
+ECU_BRANDS = [
+    'Bosch', 'Siemens', 'Delphi', 'Continental',
+    'Marelli', 'Valeo', 'Denso', 'PCR', 'EDC', 'MED',
+]
+
+READ_METHODS = [
+    "Normal Read-OBD",
+    "Virtual Read-OBD",
+    "Bench",
+    "Boot",
+]
