+++ requirements.txt
@@ -1,0 +1,1 @@
+PyQt6>=6.7.0
